Navigation Routes Changed and update fix

+++ Modules/NavigationMenu/Http/Controllers/NavigationMenuItemController.php
@@ -8,8 +8,10 @@ use Illuminate\Http\JsonResponse;
 use Modules\Core\Rules\ScopeRule;
 use Illuminate\Http\Resources\Json\JsonResource;
 use Modules\Core\Http\Controllers\BaseController;
+use Modules\NavigationMenu\Entities\NavigationMenu;
 use Illuminate\Http\Resources\Json\ResourceCollection;
 use Modules\NavigationMenu\Entities\NavigationMenuItem;
+use Modules\NavigationMenu\Exceptions\NavigationMenuItemNotFoundException;
 use Modules\NavigationMenu\Rules\NavigationMenuItemScopeRule;
 use Modules\NavigationMenu\Transformers\NavigationMenuItemResource;
 use Modules\NavigationMenu\Repositories\NavigationMenuItemRepository;
@@ -18,17 +20,18 @@ use Modules\NavigationMenu\Repositories\NavigationMenuItemValueRepository;
 class NavigationMenuItemController extends BaseController
 {
     // Protected properties
-    protected $repository, $navigation_menu_item_value_repository;
+    protected $repository, $navigation_menu, $navigation_menu_item_value_repository;
 
     /**
      * NavigationMenuItemController Class constructor
      */
-    public function __construct(NavigationMenuItemRepository $navigation_menu_item_repository, NavigationMenuItem $navigation_menu_item, NavigationMenuItemValueRepository $navigation_menu_item_value_repository)
+    public function __construct(NavigationMenuItemRepository $navigation_menu_item_repository, NavigationMenuItem $navigation_menu_item, NavigationMenu $navigation_menu, NavigationMenuItemValueRepository $navigation_menu_item_value_repository)
     {
         $this->repository = $navigation_menu_item_repository;
         $this->navigation_menu_item_value_repository = $navigation_menu_item_value_repository;
 
         $this->model = $navigation_menu_item;
+        $this->navigation_menu = $navigation_menu;
         $this->model_name = "Navigation Menu Item";
 
         // Calling Parent Constructor of BaseController
@@ -54,7 +57,7 @@ class NavigationMenuItemController extends BaseController
     /**
      * Fetches and returns the list of NavigationMenuItem
      */
-    public function index(Request $request): JsonResponse
+    public function index(Request $request, int $navigation_menu_id): JsonResponse
     {
         try
         {
@@ -62,7 +65,11 @@ class NavigationMenuItemController extends BaseController
                 "scope" => "sometimes|in:website,channel,store",
                 "scope_id" => [ "sometimes", "integer", "min:1", new ScopeRule($request->scope), new NavigationMenuItemScopeRule($request)],
             ]);
-            $fetched = $this->repository->fetchAll($request, ["values", "navigationMenu"]);
+
+            $this->navigation_menu->findOrFail($navigation_menu_id);
+            $fetched = $this->repository->fetchAll($request, ["values", "navigationMenu"], function() use($navigation_menu_id){
+                $this->model->where('navigation_menu_id', $navigation_menu_id);
+            });
         }
         catch (Exception $exception)
         {
@@ -75,11 +82,14 @@ class NavigationMenuItemController extends BaseController
     /**
      * Validates and Creates NavigationMenuItem with NavigationMenuItemValue
      */
-    public function store(Request $request): JsonResponse
+    public function store(Request $request, int $navigation_menu_id): JsonResponse
     {
         try
         {
             $data = $this->navigation_menu_item_value_repository->validateWithValuesCreate($request);
+            $this->navigation_menu->findOrFail($navigation_menu_id);
+
+            $data = array_merge($data, ['navigation_menu_id' => $navigation_menu_id]);
 
             $created = $this->repository->create($data, function ($created) use ($data) {
                 $this->navigation_menu_item_value_repository->createOrUpdate($data, $created);
@@ -96,7 +106,7 @@ class NavigationMenuItemController extends BaseController
     /**
      * Fetches and returns the NavigationMenuItem by Id
      */
-    public function show(Request $request, int $id): JsonResponse
+    public function show(Request $request, int $navigation_menu_id, int $id): JsonResponse
     {
         try
         {
@@ -106,6 +116,7 @@ class NavigationMenuItemController extends BaseController
             ]);
 
             $fetched = $this->repository->fetch($id);
+            if($fetched->navigation_menu_id !== $navigation_menu_id) throw new NavigationMenuItemNotFoundException();
         }
         catch (Exception $exception)
         {
@@ -118,13 +129,16 @@ class NavigationMenuItemController extends BaseController
     /**
      * Validates and Updates NavigationMenuItem with NavigationMenuItem values
      */
-    public function update(Request $request, int $id): JsonResponse
+    public function update(Request $request, int $navigation_menu_id, int $id): JsonResponse
     {
         try
         {
             $navigation_menu_item = $this->model->findOrFail($id);
 
             $data = $this->navigation_menu_item_value_repository->validateWithValuesUpdate($request, $navigation_menu_item);
+            $data = array_merge($data, ['navigation_menu_id' => $navigation_menu_id]);
+
+            if($navigation_menu_item->navigation_menu_id !== $navigation_menu_id) throw new NavigationMenuItemNotFoundException();
 
             $updated = $this->repository->update($data, $id, function ($updated) use ($data) {
                 $this->navigation_menu_item_value_repository->createOrUpdate($data, $updated);
@@ -142,11 +156,13 @@ class NavigationMenuItemController extends BaseController
     /**
      * Finds and Deletes NavigationMenuItem
      */
-    public function destroy(int $id): JsonResponse
+    public function destroy(int $navigation_menu_id, int $id): JsonResponse
     {
         try
         {
-            $this->model->findOrFail($id);
+            $fetched = $this->model->findOrFail($id);
+
+            if($fetched->navigation_menu_id !== $navigation_menu_id) throw new NavigationMenuItemNotFoundException();
 
             $this->repository->delete($id);
         }
@@ -161,7 +177,7 @@ class NavigationMenuItemController extends BaseController
     /**
      * Updates the Status of NavigationMenuItem with given Id
      */
-    public function updateStatus(Request $request, int $id): JsonResponse
+    public function updateStatus(Request $request, int $navigation_menu_id, int $id): JsonResponse
     {
         try
         {
@@ -178,7 +194,7 @@ class NavigationMenuItemController extends BaseController
     /**
      * Fetches and returns Attributes for NavigationMenuItem Values
      */
-    public function attributes(Request $request): JsonResponse
+    public function attributes(Request $request, int $navigation_menu_id): JsonResponse
     {
         try
         {
